fix group restrictions per page view for cms
