+++ apcd-cms/src/taccsite_cms/templatetags/limit_visibility_in_menu.py
@@ -1,18 +1,8 @@
 from django import template
-from urllib.parse import urlparse
-from django.utils.html import format_html
 from django.contrib import auth as auth
+from apps.utils.apcd_groups import has_apcd_group, is_apcd_admin
 
 register = template.Library()
-
-def has_apcd_group(user):
-    for group in user.groups.all():
-        if group.name in ['APCD_ADMIN', 'SUBMITTER_ADMIN', 'SUBMITTER_USER']:
-            return True
-    return False
-
-def is_apcd_admin(user):
-    return user.groups.filter(name='APCD_ADMIN').exists()
 
 @register.simple_tag(takes_context=True)
 def limit_visibility_in_menu(context, menu_item):
